+++ binance/src/main/java/com/s8/binance/model/request/CoinRequestDto.java
@@ -1,7 +1,6 @@
 package com.s8.binance.model.request;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -26,5 +25,5 @@ public class CoinRequestDto implements Serializable {
     private String description;
 
     @ApiModelProperty(example = "27661.90")
-    private BigDecimal usdValue;
+    private Double usdValue;
 }
